Add new buffer control class.

+++ includes/TerreateGraphics.hpp
@@ -1,16 +1,17 @@
 #ifndef __TERREATE_GRAPHICS_TERREATEGRAPHICS_HPP__
 #define __TERREATE_GRAPHICS_TERREATEGRAPHICS_HPP__
 
-#include "animation.hpp"
+/* #include "animation.hpp" */
 #include "buffer.hpp"
 #include "defines.hpp"
 #include "event.hpp"
 #include "font.hpp"
 #include "job.hpp"
 #include "model.hpp"
+#include "newbuffer.hpp"
 #include "screen.hpp"
 #include "shader.hpp"
-#include "skeleton.hpp"
+/* #include "skeleton.hpp" */
 #include "text.hpp"
 #include "texture.hpp"
 #include "window.hpp"
@@ -31,7 +32,7 @@ public:
   void Frame(Uint const &fps);
 
 public:
-  static Float GetCurrentRuntime() { return glfwGetTime(); }
+  static Double GetCurrentRuntime() { return glfwGetTime(); }
 };
 
 void Initialize();
